feat: Add pagination to task list
As the list of tasks grows, displaying all of them on a single page is
inefficient and leads to a poor user experience.
This change introduces a basic pagination system to show a limited
number of tasks per page.
- Implements pagination via a URL query parameter (`?page=N`)
- The number of tasks per page is controlled by the `PER_PAGE` constant
- Redirects users to page 1 if a requested page is too low or to the
last page if it's too high
- Passes pagination-related variables
(`has_next`, `has_prev`, `total_pages`)
to the template to enable rendering navigation links